Pass Event::PressMove for mouse input immediately
This is a partial revert of 2d8e68f3 which has been half
broken since due to frequent clearing of pending events.
It's also the wrong solution since it decreases precision of
momentum scrolling algorithms.
